Use parameter file common.pf, fix month off-by-one
Code is now using the common.pf from the web site, so the script needs
to have the PFPATH environment variable set correctly.

Month was using the current month rather than the previous month. Since
the monthly images are generated for a given month at the beginning of
the following month, the file naming is now set to match that.

Removed references to symlinks - everything is now about generating
stills in JPEG format.